Back out XrmQuark change.

+++ lib/X11/Xresource.h
@@ -42,7 +42,7 @@ ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS
 SOFTWARE.
 
 ******************************************************************/
-/* $XFree86: xc/lib/X11/Xresource.h,v 3.2 1998/10/03 08:41:30 dawes Exp $ */
+/* $XFree86: xc/lib/X11/Xresource.h,v 3.3 2000/09/28 14:50:55 tsi Exp $ */
 
 #ifndef _XRESOURCE_H_
 #define _XRESOURCE_H_
@@ -79,7 +79,7 @@ extern char *Xpermalloc(
  *
  ****************************************************************/
 
-typedef long     XrmQuark, *XrmQuarkList;
+typedef int       XrmQuark, *XrmQuarkList;
 #define NULLQUARK ((XrmQuark) 0)
 
 typedef char *XrmString;
